requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
opencv-python
scikit-learn
matplotlib
seaborn
pandas
